arch: Create a common standard C++ time implementation
