+++ src/arch/macos/CMakeLists.txt
@@ -26,7 +26,7 @@ if("${FORTE_ARCHITECTURE}" STREQUAL "MacOs")
   else(FORTE_FAKE_TIME)
     forte_set_timer(../posix/pctimeha)
   endif(FORTE_FAKE_TIME)
-  forte_add_sourcefile_cpp(../posix/forte_architecture_time.cpp)
+  forte_add_sourcefile_cpp(../posix/forte_architecture_time.cpp ../forte_standard_time.cpp)
 
   forte_add_sourcefile_hcpp(../posix/forte_thread forte_sync forte_sem)
   forte_add_sourcefile_cpp(../genforte_printer.cpp)
