+++ src/arch/posix/forte_architecture_time.cpp
@@ -1,6 +1,7 @@
 /*******************************************************************************
  * Copyright (c) 2019 TU Wien/ACIN
  *               2023 Martin Erich Jobst
+ *               2024 Samator Indo Gas
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
@@ -10,28 +11,11 @@
  * Contributors:
  *  Martin Melik Merkumians - Adds getNanoSecondsMonotonic
  *  Martin Jobst - add high-resolution realtime clock
+ *  Ketut Kumajaya - Use the standard C++ time implementation
  *******************************************************************************/
-
-#include <time.h>
 
 #include "forte_architecture_time.h"
 #include "forte_constants.h"
-
-#ifndef FORTE_FAKE_TIME
-uint_fast64_t getNanoSecondsMonotonic() {
-  struct timespec now;
-  clock_gettime(CLOCK_MONOTONIC, &now);
-  return static_cast<uint_fast64_t>(now.tv_nsec) + static_cast<uint_fast64_t>(now.tv_sec) *
-   static_cast<uint_fast64_t>(forte::core::constants::cNanosecondsPerSecond);
-}
-
-uint_fast64_t getNanoSecondsRealtime() {
-  struct timespec now;
-  clock_gettime(CLOCK_REALTIME, &now);
-  return static_cast<uint_fast64_t>(now.tv_nsec) + static_cast<uint_fast64_t>(now.tv_sec) *
-   static_cast<uint_fast64_t>(forte::core::constants::cNanosecondsPerSecond);
-}
-#endif
 
 time_t forte_timegm(struct tm *pa_tm) {
   return timegm(pa_tm);
