+++ src/arch/win32/forte_architecture_time.cpp
@@ -15,24 +15,12 @@
  *  Martin Jobst - add high-resolution realtime clock
  *  Ketut Kumajaya - Use the standard C++ time functions, fix pre Windows 8/
  *                   Server 2012 compatibility issue
+ *                 - Move the standard C++ time functions as a common time
+ *                   implementation
  *******************************************************************************/
-
-#include <chrono>
-
-using namespace std::chrono;
 
 #include "forte_architecture_time.h"
 #include "forte_constants.h"
-
-#ifndef FORTE_FAKE_TIME
-uint_fast64_t getNanoSecondsMonotonic() {
-  return static_cast<uint_fast64_t>(duration_cast<nanoseconds>(steady_clock::now().time_since_epoch()).count());
-}
-
-uint_fast64_t getNanoSecondsRealtime() {
-  return static_cast<uint_fast64_t>(duration_cast<nanoseconds>(system_clock::now().time_since_epoch()).count());
-}
-#endif
 
 time_t forte_timegm(struct tm *pa_tm) {
   return _mkgmtime(pa_tm);
